// frontend/src/client/models/AddressUpdate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Address update schema
 */
export type AddressUpdate = {
    name?: (string | null);
    receiver_name?: (string | null);
    receiver_phone?: (string | null);
    address_line?: (string | null);
    ward?: (string | null);
    district?: (string | null);
    city?: (string | null);
    postal_code?: (string | null);
    is_default?: (boolean | null);
    notes?: (string | null);
};